[cli] moves sources under a src directory
