--- bin/mscgen_js
@@ -1,2 +1,0 @@
-#!/usr/bin/env node
-require("../src/script/cli/mscgen.js");
