+++ src/script/cli/bin/mscgen_js
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+require("../src/mscgen.js");
