+++ src/script/cli/src/cli-phantom-vector.js
