+++ src/script/cli/src/normalizations.js
