+++ src/script/cli/src/validations.js
